Replace Quick/Nimble with XCTest
Using XCTest provides better integration with Xcode. For instance,
with Quick you cannot run a single test or jump to a test by selecting
it in the Test Navigator. There's also a bunch of other issues, like
not getting a proper stacktrace when setting breakpoints inside tests.

I find tests to be a lot more readable when using Quick, but it
doesn't weight up for all the issues. It's also nice to remove any
third-party dependencies from Parchment, even if they are just for
development purposes.
