--- Cartfile.resolved
@@ -1,2 +1,0 @@
-github "Quick/Nimble" "v8.0.2"
-github "Quick/Quick" "v2.1.0"
