+++ ParchmentTests/Item.swift
@@ -16,3 +16,11 @@ func ==(lhs: Item, rhs: Item) -> Bool {
 func <(lhs: Item, rhs: Item) -> Bool {
   return lhs.index < rhs.index
 }
+
+final class ItemCell: PagingCell {
+  private(set) var item: Item?
+  
+  override func setPagingItem(_ pagingItem: PagingItem, selected: Bool, options: PagingOptions) {
+    self.item = pagingItem as? Item
+  }
+}
